refactor(eval): split event.rs into module directory
